Move collectionParams out of I19 and somewhere more general
